adding RangedEntrySlider for later use

+++ common/src/main/java/net/xstopho/resourceconfigapi/annotations/RangedEntry.java
@@ -6,6 +6,6 @@ import java.lang.annotation.RetentionPolicy;
 @Retention(RetentionPolicy.RUNTIME)
 public @interface RangedEntry {
 
-    float minValue();
-    float maxValue();
+    double minValue();
+    double maxValue();
 }
